class-projects/Big Agents/Game Master Agent/pyproject.toml
[project]
name = "game-master-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
]
